refactor: DataLoader 를 test 용과 prod 용으로 분리합니다
